Make symmetric the streaming of enums
m_fieldId is a QContactDetail::DetailType and it is decoded with a
quint32:

  quint32 type;
  stream >> type >> m_fieldId >> m_minValue >> m_maxValue >> flags >> rangeFlags;

+++ src/contacts/filters/qcontactdetailrangefilter_p.h
@@ -94,7 +94,7 @@ public:
     QDataStream& outputToStream(QDataStream& stream, quint8 formatVersion) const
     {
         if (formatVersion == 1) {
-            stream << m_typeId << m_fieldId << m_minValue << m_maxValue
+            stream << quint32(m_typeId) << m_fieldId << m_minValue << m_maxValue
                 << static_cast<quint32>(m_flags)
                 << static_cast<quint32>(m_rangeflags);
         }
